backend/scripts/requirements.txt
pymongo==4.7.2
python-dotenv==1.1.0
python_bcrypt==0.3.2
